pyproject.toml
[project]
name = "python-daft"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "getdaft>=0.4.1",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "ray[data]>=2.40.0",
]
